+++ calva/repl-window.ts
@@ -1,10 +1,9 @@
 import * as vscode from "vscode";
 import { cljSession, cljsSession } from "./connector"
 import * as path from "path";
-import * as fs from "fs";
 import * as state from "./state";
 import status from "./status"
-import { readFileSync } from "fs";
+import * as fs from "fs";
 import { NReplEvaluation, NReplSession } from "./nrepl";
 
 import annotations from './providers/annotations';
@@ -23,6 +22,13 @@ export function activeReplWindow() {
     return undefined;
 }
 
+export function isReplWindowOpen(mode: "clj" | "cljs" = "clj") {
+    // If we find `mode` in ythe `replWindows` dictionary, then it is open.
+    if (!replWindows[mode]) {
+        return(false);   
+    }
+    return(true);
+}
 
 class REPLWindow {
     evaluation: NReplEvaluation;
@@ -96,10 +102,10 @@ class REPLWindow {
 
         panel.onDidChangeViewState(e => {
             this.useBuffer = !e.webviewPanel.visible;
-
             if (e.webviewPanel.visible) {
                 this.buffer.forEach(x => this.panel.webview.postMessage(x))
                 this.buffer = [];
+                replViewColum[this.type] = e.webviewPanel.viewColumn;
             }
             status.update();
         })
@@ -108,18 +114,17 @@ class REPLWindow {
         // TODO: Add a custom-cljs.svg
         const cljTypeSlug = `clj-type-${cljType.replace(/ /, "-").toLowerCase()}`;
         const cljsTypeSlug = `cljs-type-${cljsType.replace(/ /, "-").toLowerCase()}`;
-        let html = readFileSync(path.join(ctx.extensionPath, "html/index.html")).toString();
-        html = html.replace("{{baseUri}}", getUrl());
-        html = html.replace("{{script}}", getUrl("/main.js"));
-        html = html.replace("{{font}}", getUrl("/fira_code.css"));
-        html = html.replace("{{logo-symbol}}", getUrl(`/images/calva-symbol-logo.svg`));
+        let html = fs.readFileSync(path.join(ctx.extensionPath, "html/index.html")).toString();
+        let script = vscode.Uri.file(path.join(ctx.extensionPath, "html/main.js")).with({ scheme: 'vscode-resource' }).toString()
+        html = html.replace("{{script}}", script);
+        html = html.replace("{{logo-symbol}}", getImageUrl(`calva-symbol-logo.svg`));
         html = html.replace(/{{hero-classes}}/g, `${type} ${cljTypeSlug} ${cljsTypeSlug}`);
         html = html.replace("{{clj-type}}", `${cljType.replace(/ /g, "&nbsp;")}`);
+        html = html.replace("{{clj-type-logo}}", getImageUrl(`${cljTypeSlug}.svg`));
+        html = html.replace("{{clj-logo}}", getImageUrl(`clj.svg`));
         html = html.replace("{{cljs-type}}", `${cljsType.replace(/ /g, "&nbsp;")}`);
-        html = html.replace("{{clj-type-logo}}", getUrl(`/images/${cljTypeSlug}.svg`));
-        html = html.replace("{{clj-logo}}", getUrl(`/images/clj.svg`));
-        html = html.replace("{{cljs-type-logo}}", getUrl((`/images/${cljsTypeSlug}.svg`)));
-        html = html.replace("{{cljs-logo}}", getUrl(`/images/cljs.svg`));
+        html = html.replace("{{cljs-type-logo}}", getImageUrl((`${cljsTypeSlug}.svg`)));
+        html = html.replace("{{cljs-logo}}", getImageUrl(`cljs.svg`));
         panel.webview.html = html;
 
         this.connect().catch(reason => {
@@ -169,6 +174,10 @@ class REPLWindow {
         })
         try {
             this.postMessage({ type: "repl-response", value: await this.evaluation.value, ns: this.ns = ns || this.evaluation.ns || this.ns });
+            if(this.evaluation.ns && this.ns != this.evaluation.ns) {
+                // the evaluation changed the namespace so set the new namespace.
+                this.setNamespace(this.evaluation.ns);
+            }
         } catch (e) {
             this.postMessage({ type: "repl-error", ex: e });
             let stacktrace = await this.session.stacktrace();
@@ -189,12 +198,20 @@ class REPLWindow {
 let ctx: vscode.ExtensionContext
 
 let replWindows: { [id: string]: REPLWindow } = {};
+let replViewColum: { [id: string]: vscode.ViewColumn } = {"clj": vscode.ViewColumn.Two, 
+                                                          "cljs": vscode.ViewColumn.Two};
 
-function getUrl(name?: string) {
-    if (name)
-        return vscode.Uri.file(path.join(ctx.extensionPath, "html", name)).with({ scheme: 'vscode-resource' }).toString()
+function getImageUrl(name: string) {
+    let imagepath = "";
+    if (!name)
+         imagepath = path.join(ctx.extensionPath, "html/images/empty.svg");
     else
-        return vscode.Uri.file(path.join(ctx.extensionPath, "html")).with({ scheme: 'vscode-resource' }).toString()
+         imagepath = path.join(ctx.extensionPath, "html/images/", name);
+
+    if(!fs.existsSync(imagepath)) {
+        imagepath = path.join(ctx.extensionPath, "html/images/empty.svg");
+    }
+    return vscode.Uri.file(imagepath).with({ scheme: 'vscode-resource' }).toString()
 }
 
 export async function reconnectReplWindow(mode: "clj" | "cljs") {
@@ -202,6 +219,26 @@ export async function reconnectReplWindow(mode: "clj" | "cljs") {
         await replWindows[mode].connect()
         replWindows[mode].postMessage({ type: "reconnected", ns: replWindows[mode].ns });
     }
+}
+
+export async function openClojureReplWindows() {
+    if (state.deref().get('connected')) {
+        if(util.getSession("clj")) {
+            openReplWindow("clj", true);
+            return;
+        }
+    }
+    vscode.window.showInformationMessage("Not connected to a Clojure REPL server");
+}
+
+export async function openClojureScriptReplWindows() {
+    if (state.deref().get('connected')) {
+        if(util.getSession("cljs")) {
+            openReplWindow("cljs", true);
+            return;
+        }
+    }
+    vscode.window.showInformationMessage("Not connected to a ClojureScript REPL server");
 }
 
 export async function openReplWindow(mode: "clj" | "cljs" = "clj", preserveFocus: boolean = true) {
@@ -214,7 +251,7 @@ export async function openReplWindow(mode: "clj" | "cljs" = "clj", preserveFocus
         replWindows[mode].session = await session.clone();
     }
 
-    replWindows[mode].panel.reveal(vscode.ViewColumn.Two, preserveFocus);
+    replWindows[mode].panel.reveal(replViewColum[mode], preserveFocus);
     return replWindows[mode];
 }
 
@@ -232,7 +269,7 @@ export async function createReplWindow(session: NReplSession, mode: "clj" | "clj
     const sessionClone = await session.clone();
     let title = mode == "clj" ? "CLJ REPL" : "CLJS REPL";
     const panel = vscode.window.createWebviewPanel("replInteractor", title, {
-        viewColumn: vscode.ViewColumn.Two,
+        viewColumn: replViewColum[mode],
         preserveFocus: true
     }, {
         retainContextWhenHidden: true,
@@ -246,13 +283,13 @@ export async function createReplWindow(session: NReplSession, mode: "clj" | "clj
 }
 
 async function loadNamespaceCommand(reload = true) {
+    await openReplWindow(util.getREPLSessionType(), reload);
     await setREPLNamespace(util.getDocumentNamespace(), reload).catch(r => { console.error(r) });
-    await openReplWindow(util.getREPLSessionType());
 }
 
 async function setREPLNamespaceCommand() {
-    await setREPLNamespace(util.getDocumentNamespace(), false).catch(r => { console.error(r) });
     await openReplWindow(util.getREPLSessionType());
+    await setREPLNamespace(util.getDocumentNamespace(), false).catch(r => { console.error(r) });
 }
 
 export async function sendTextToREPLWindow(text, ns: string, pprint: boolean) {
@@ -294,7 +331,7 @@ function evalCurrentFormInREPLWindow(topLevel: boolean, pprint: boolean) {
 
     if (selection.isEmpty) {
         codeSelection = select.getFormSelection(doc, selection.active, topLevel);
-        annotations.decorateSelection(codeSelection, editor, annotations.AnnotationStatus.REPL_WINDOW);
+        annotations.decorateSelection("", codeSelection, editor, annotations.AnnotationStatus.REPL_WINDOW);
         code = doc.getText(codeSelection);
     } else {
         codeSelection = selection;
@@ -315,8 +352,8 @@ function evalCurrentTopLevelFormInREPLWindowCommand() {
 
 export function activate(context: vscode.ExtensionContext) {
     ctx = context;
-    context.subscriptions.push(vscode.commands.registerCommand('calva.openCljReplWindow', () => openReplWindow("clj", true)));
-    context.subscriptions.push(vscode.commands.registerCommand('calva.openCljsReplWindow', () => openReplWindow("cljs"), true));
+    context.subscriptions.push(vscode.commands.registerCommand('calva.openCljReplWindow', openClojureReplWindows));
+    context.subscriptions.push(vscode.commands.registerCommand('calva.openCljsReplWindow', openClojureScriptReplWindows));
     context.subscriptions.push(vscode.commands.registerCommand('calva.loadNamespace', loadNamespaceCommand));
     context.subscriptions.push(vscode.commands.registerCommand('calva.setREPLNamespace', setREPLNamespaceCommand));
     context.subscriptions.push(vscode.commands.registerCommand('calva.evalCurrentFormInREPLWindow', evalCurrentFormInREPLWindowCommand));